+++ src/utils/pr.ts
@@ -3,18 +3,7 @@ import * as action from "./action.ts";
 import { context } from "./context.ts";
 import { RestEndpointMethodTypes } from "../deps.ts";
 
-export async function lock() {
-  const prNumber = context.issue.number;
-  await octokit.issues.lock({
-    ...context.repo,
-    issue_number: prNumber,
-  }).catch((error) => {
-    throw new Error(
-      `Error occurred when locking the pull request #${prNumber}: ${error.message}`,
-    );
-  });
-  action.info(`Successfully locked the pull request #${prNumber}`);
-}
+export { lock } from "./issue.ts";
 
 export async function branch(): Promise<string> {
   const prNumber = context.issue.number;
